+++ R/get_output_data.R
@@ -14,9 +14,10 @@
 #' get_output_data(c("192.168.0.12", "192.168.0.230"))
 #' }
 #'
-#' @importFrom dplyr mutate across ends_with rename select ends_with
+#' @importFrom dplyr mutate across ends_with rename select everything
 #' @importFrom tidyr pivot_longer separate_wider_regex pivot_wider
 #' @importFrom purrr map_dfr
+#' @importFrom lubridate now
 #' @importFrom units set_units
 #' @importFrom rlang .data
 #'
@@ -24,7 +25,7 @@ get_output_data <- function(device_ip, model = "APSystems", ...) {
 
   if (model == "APSystems") {
     out_tbl <- query_ap_devices(device_ip, "getOutputData") |>
-      pivot_longer(!device_id) |>
+      pivot_longer(!"device_id") |>
       separate_wider_regex("name", patterns = c("metric" = "\\D+","inverter" = "\\d+")) |>
       pivot_wider(names_from = "metric", values_from = "value")|>
       rename(output_power = "p", today_energy = "e", lifetime_energy = "te")
@@ -34,15 +35,30 @@ get_output_data <- function(device_ip, model = "APSystems", ...) {
     )
 
   } else if (model == "Enphase-Envoy-S") {
-    out_tbl <- map_dfr(device_ip, ~query_enphase_device(.x, "production/inverters/") |>
+    out_tbl <- map_dfr(device_ip, ~query_enphaseenvoy_device(.x, "production/inverters/") |>
       rename(output_power = "lastReportWatts", output_max_power = "maxReportWatts",
              last_report = "lastReportDate"
       ))
     mutate(out_tbl,
-           last_report = as.POSIXct(last_report),
+           last_report = as.POSIXct(.data$last_report),
            # TODO BUG may fail if not parsed as number
            across(ends_with("_power"), \(x) set_units(x, "W"))
     )
+
+  } else if (model == "Enphase-Energy") {
+    out_tbl <- map_dfr(device_ip, ~query_enphaseenergy_device(.x, "stream/meter") |>
+                         pivot_longer(!"device_id") |>
+                         separate_wider_regex(name, patterns = c(name = ".*tion", ".", phase = "ph\\.\\w", ".", metric = "\\w$")) |>
+                         pivot_wider(names_from = "metric", values_from = "value") |>
+                         rename(power = "p", voltage = "v", current = "i")
+                       )
+    mutate(out_tbl,
+           last_report = now(),
+           across(ends_with("power"), \(x) set_units(x, "W")),
+           across(ends_with("voltage"), \(x) set_units(x, "V")),
+           across(ends_with("current"), \(x) set_units(x, "A"))
+    ) |>
+    select("device_id", "last_report", everything())
 
   } else if (model == "Fronius") {
     out_tbl <- query_fronius_devices(device_ip, "GetInverterRealtimeData.cgi?Scope=System") |>
@@ -51,7 +67,7 @@ get_output_data <- function(device_ip, model = "APSystems", ...) {
       ) |>
       select(-ends_with(".Unit"))
     mutate(out_tbl,
-           last_report = as.POSIXct(last_report),
+           last_report = as.POSIXct(.data$last_report),
            across(ends_with("_power"), \(x) set_units(x, "W")),
            across(ends_with("_energy"), \(x) set_units(x, "kW/h"))
     )